controller: move rollout code to rollout.go
Move driveRollout(), nodeState enum, classifyNode(), and
TestClassifyNode to dedicated rollout.go and rollout_test.go files.
The controller file is already ~500 lines and the upcoming rollout
state machine commits will add substantially more. Separating now
keeps the split clean: bootcnodepool_controller.go owns Reconcile(),
watches, and membership sync; rollout.go owns the rollout state
machine.

Pure code move, no behavioral changes.
